Codecs/FLAC: Decode 32-bit streams (RFC 9639 bit depth code 0b111)
The frame parser treated bit depth code 0b111 as reserved, rejecting
every frame of a 32-bit FLAC file even though the demuxer and the
FLACRFC9639 helpers map it correctly to 32 bits (RFC 9639 Section
9.1.4, Table 17). Decoding 32-bit streams also requires 33-bit side
subframe samples (Section 9.2.2), which could not be read or stored:

- BitstreamReader: add readBitsSigned64 for sample reads above 32 bits
- SubframeDecoder/ChannelDecorrelator: widen sample buffers to int64
  so 33-bit side-channel warm-up, verbatim, constant, and predicted
  samples survive prediction and decorrelation unclipped
- NativeFLACCodec: decode into int64 buffers, then narrow (with
  defensive clamping) to int32 after decorrelation for the sample
  reconstructor and MD5 validator, whose values always fit 32 bits

+++ include/codecs/flac/ChannelDecorrelator.h
@@ -57,7 +57,7 @@ public:
      * 
      * Requirements: 7.1-7.8, 26.1-26.8
      */
-    bool decorrelate(int32_t** channels, uint32_t block_size,
+    bool decorrelate(int64_t** channels, uint32_t block_size,
                     uint32_t channel_count, ChannelAssignment assignment);
     
 private:
@@ -71,7 +71,7 @@ private:
      * 
      * Requirements: 7.2
      */
-    void decorrelateLeftSide(int32_t* left, int32_t* side, uint32_t count);
+    void decorrelateLeftSide(int64_t* left, int64_t* side, uint32_t count);
     
     /**
      * Decorrelate right-side stereo
@@ -83,7 +83,7 @@ private:
      * 
      * Requirements: 7.3
      */
-    void decorrelateRightSide(int32_t* side, int32_t* right, uint32_t count);
+    void decorrelateRightSide(int64_t* side, int64_t* right, uint32_t count);
     
     /**
      * Decorrelate mid-side stereo
@@ -98,7 +98,7 @@ private:
      * 
      * Requirements: 7.4, 7.5
      */
-    void decorrelateMidSide(int32_t* mid, int32_t* side, uint32_t count);
+    void decorrelateMidSide(int64_t* mid, int64_t* side, uint32_t count);
     
     /**
      * Validate channel count
